% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_uniform_patches.R
\name{generate_uniform_patches}
\alias{generate_uniform_patches}
\title{Generate Uniformly Distributed Patches}
\usage{
generate_uniform_patches(n.patches, extent)
}
\arguments{
\item{n.patches}{Integer. The number of habitat patches to generate.}

\item{extent}{Numeric. The side length of the square landscape in which patches are placed.}
}
\value{
A list with numeric vectors \code{x} and \code{y}, representing the coordinates of each patch.
}
\description{
Creates a spatially uniform distribution of habitat patches within a square landscape.
}
\examples{
patches <- generate_uniform_patches(n.patches = 100, extent = 10)

}
